requirements.txt
django>=2.2.0,<2.2.99
djangorestframework>=3.9.0,<3.9.99
pillow>=3.3.0,<3.4.99
requests==2.9.1
redis
pytz
tzlocal
jdatetime
django-debug-toolbar
django_rest_framework_redocs
